+++ server/src/main/java/com/bsds/server/repository/ResortRepository.java
@@ -1,4 +1,4 @@
-package com.bsds.server;
+package com.bsds.server.repository;
 
 import com.bsds.server.db.ResortEntity;
