+++ server/src/main/java/com/bsds/server/repository/SkierRepository.java
@@ -1,4 +1,4 @@
-package com.bsds.server;
+package com.bsds.server.repository;
 
 import com.bsds.server.db.SkierEntity;
